Added Details app and updated Atlas app appropriately
Added the new Details app to handle the logic for anything having
details. This required some changes to be made to Atlas as it was
originally trying to do this in itself, but was failing. Updated the
relevant admin pages for easier updating. Updated the altas views to
include parents, children, and details.