Added Navbar
Added a Navbar that will get its menu entries from each app. Each app will that wants menu entries will now maintain a nav_templates value in apps.py that will point to its portion of the navbar. A context processor in the dndsite will iterate over the apps and collect all these templates to be included in the main navbar. The navbar was made part of the base.html page.