Added Endeavours App
The Endeavours app will track quests and other things that the party undertakes.
This is currently just a skeleton of the class to hold the name of quests and relate to them. It is intended later to flesh out this section with steps in a quest, and possibly also things like quest types.